requirements.txt
asgiref>=3.4.1
bcrypt
cffi
click==8.0.3
colorama
cryptography
Django>=3.2,<3.3
django-appconf
mssql-django>=1.1.1,<2
greenlet
Jinja2>=3
MarkupSafe>=1.1
paramiko>=2.8.1
python-win-ad
python-dateutil>=2.8.1
pytz
pywin32>=300
sqlparse>=0.4.2
toml>=0.10.2
wrapt>=1.11,<1.13
fuzzywuzzy
djangorestframework>=3.13.1,<4
django-filter>=21.1
drf-yasg==1.20.0
django-mptt>=0.13.4
sphinx<5,>=3.1
myst-parser>=0.17
sphinx_rtd_theme
black
python-Levenshtein